A4A: register the agency site routes as their own route group
Moves the agency site routes out of the agency pages layout into a
createAgencySitesRoutes() group behind a new supports.agencySites flag,
per the direction agreed in the shared-site-tree discussion. The flag
replaces supports.agency.sites, so one setting controls both the routes
and the sidebar entry. The agency guard (client-user block and team
capability checks) is shared by both layouts. No URL changes.
