+++ client/dashboard/agency/sites/site/performance.tsx
@@ -1,4 +1,4 @@
-import { agencySiteRoute } from '../../../app/router/agency';
+import { agencySiteRoute } from '../../../app/router/agency-sites';
 import { SitePerformanceContent } from '../../../sites/performance';
 
 export default function AgencySitePerformance() {
