+++ client/dashboard/agency/sites/site/scan-page.tsx
@@ -1,7 +1,7 @@
 import { siteBySlugQuery } from '@automattic/api-queries';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { useSiteTimezone } from '../../../app/hooks/use-site-timezone';
-import { agencySiteRoute } from '../../../app/router/agency';
+import { agencySiteRoute } from '../../../app/router/agency-sites';
 import { ScanContent } from '../../../sites/scan/scan-content';
 
 export default function AgencySiteScanPage( { scanTab }: { scanTab: 'active' | 'history' } ) {
