+++ client/dashboard/app-a4a/index.tsx
@@ -27,10 +27,10 @@ boot( {
 			exclusiveOffers: true,
 			learn: true,
 			mcp: true,
-			sites: true,
 			team: true,
 			earn: true,
 		},
+		agencySites: true,
 		agencyClient: { subscriptions: true },
 		sites: false,
 		domains: false,
