+++ client/dashboard/app-ciab/index.tsx
@@ -26,6 +26,7 @@ boot( {
 	LoadingLogo: CiabDashboardStepperLogo,
 	supports: {
 		agency: false,
+		agencySites: false,
 		agencyClient: false,
 		sites: true,
 		domains: true,
